+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v2.7.3 - November 13, 2025
+
+Fixes:
+
+- Find latest VsDevCmd.bat correctly when Visual Studio 2026 is installed.
+- Fix issue where if only a single Visual Studio instance is installed, an error is thrown when trying to find the latest MSBuild.exe file.
+
 ## v2.7.2 - November 12, 2025
 
 Fixes: